Move transaction list item into its own component
